app completed

+++ lib/feature/tasks/widgets/task_tile.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_slidable/flutter_slidable.dart';
 import 'package:provider/provider.dart';
 import '../../../data/model/task_model.dart';
 import '../controller/task_controller.dart';
+import '../view/home_screen.dart';
 
 class TaskTile extends StatelessWidget {
   final TaskModel task;
@@ -9,31 +11,58 @@ class TaskTile extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Card(
-      margin: const EdgeInsets.symmetric(horizontal: 12, vertical: 6),
-      child: ListTile(
-        onTap: () {
-          Navigator.pushNamed(context, "/add", arguments: task);
-        },
-        leading: Checkbox(
-          value: task.isCompleted,
-          onChanged: (_) =>
-              context.read<TaskController>().toggleTask(task),
-        ),
-        title: Text(
-          task.title,
-          style: TextStyle(
-            decoration:
-            task.isCompleted ? TextDecoration.lineThrough : null,
+    return Slidable(
+      key: ValueKey(task.id),
+      endActionPane: ActionPane(
+        motion: const DrawerMotion(),
+        children: [
+          SlidableAction(
+            onPressed: (_) async {
+              final ok = await showConfirmDialog(
+                context,
+                "Delete this task?",
+              );
+
+              if (ok) {
+                context.read<TaskController>().deleteTask(task.id);
+              }
+            },
+            icon: Icons.delete,
+            backgroundColor: Colors.red,
+            label: 'Delete',
           ),
-        ),
-        subtitle: Text(
-          "${task.priority}${task.dueDate != null ? " • ${task.dueDate!.toLocal().toString().split(" ")[0]}" : ""}",
-        ),
-        trailing: IconButton(
-          icon: const Icon(Icons.delete),
-          onPressed: () =>
-              context.read<TaskController>().deleteTask(task.id),
+        ],
+      ),
+      child: Card(
+        margin: const EdgeInsets.symmetric(horizontal: 12, vertical: 6),
+        elevation: 2,
+        child: ListTile(
+          leading: Checkbox(
+            value: task.isCompleted,
+            onChanged: (_) =>
+                context.read<TaskController>().toggleTask(task),
+          ),
+          title: Text(
+            task.title,
+            style: TextStyle(
+              fontWeight: FontWeight.w600,
+              decoration: task.isCompleted ? TextDecoration.lineThrough : null,
+            ),
+            maxLines: 1,
+            overflow: TextOverflow.ellipsis,
+          ),
+          subtitle: Text(
+            "${task.priority}${task.dueDate != null ? " • ${task.dueDate!.toLocal().toString().split(" ")[0]}" : ""}",
+          ),
+          trailing: GestureDetector(
+            onTap: () {
+              Navigator.pushNamed(context, "/add", arguments: task);
+            },
+            child: const Icon(
+              Icons.edit,
+              size: 30.0,
+            ),
+          ),
         ),
       ),
     );
